Revert "change of api routes for deployment"
This reverts commit b5065e3a1e8ce85195721cb895938767ce39a3e8.
