+++ sbrp-frontend/api/skills/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+skills = Blueprint('skills', __name__)
+
+from . import views
